requirements.txt
Flask==3.0.3
requests==2.31.0
python-dotenv==1.0.1
groq==0.5.0
networkx==3.2.1
flask-sqlalchemy
werkzeug